// generated/graphql-storefront.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K]
}
// ...
export type Scalars = {
	ID: { input: string; output: string }
	String: { input: string; output: string }
	Boolean: { input: boolean; output: boolean }
	Int: { input: number; output: number }
	Float: { input: number; output: number }
	Color: { input: any; output: any }
	DateTime: { input: any; output: any }
	Decimal: { input: any; output: any }
	HTML: { input: any; output: any }
	JSON: { input: any; output: any }
	URL: { input: any; output: any }
	UnsignedInt64: { input: any; output: any }
}
// ...
export type Article = HasMetafields &
	Node &
	OnlineStorePublishable &
	Trackable & {
		__typename?: "Article"
		/**
		 * The article's author.
		 * @deprecated Use `authorV2` instead.
		 */
		author: ArticleAuthor
		/** The article's author. */
		authorV2?: Maybe<ArticleAuthor>
		/** The blog that the article belongs to. */
		blog: Blog
		/** List of comments posted on the article. */
		comments: CommentConnection
		/** Stripped content of the article, single line with HTML tags removed. */
		content: Scalars["String"]["output"]
		/** The content of the article, complete with HTML formatting. */
		contentHtml: Scalars["HTML"]["output"]
		/** Stripped excerpt of the article, single line with HTML tags removed. */
		excerpt?: Maybe<Scalars["String"]["output"]>
		/** The excerpt of the article, complete with HTML formatting. */
		excerptHtml?: Maybe<Scalars["HTML"]["output"]>
		/** A human-friendly unique string for the Article automatically generated from its title. */
		handle: Scalars["String"]["output"]
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** The image associated with the article. */
		image?: Maybe<Image>
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars["URL"]["output"]>
		/** The date and time when the article was published. */
		publishedAt: Scalars["DateTime"]["output"]
		/** The article’s SEO information. */
		seo?: Maybe<Seo>
		/**
		 * A categorization that a article can be tagged with.
		 *
		 */
		tags: Array<Scalars["String"]["output"]>
		/** The article’s name. */
		title: Scalars["String"]["output"]
		/** A URL parameters to be added to a page URL when it is linked from a GraphQL result. This allows for tracking the origin of the traffic. */
		trackingParameters?: Maybe<Scalars["String"]["output"]>
	}
// ...
export type ArticleAuthor = {
	__typename?: "ArticleAuthor"
	/** The author's bio. */
	bio?: Maybe<Scalars["String"]["output"]>
	/** The author’s email. */
	email: Scalars["String"]["output"]
	/** The author's first name. */
	firstName: Scalars["String"]["output"]
	/** The author's last name. */
	lastName: Scalars["String"]["output"]
	/** The author's full name. */
	name: Scalars["String"]["output"]
}

/**
 * An auto-generated type for paginating through multiple Articles.
 *
 */
export type ArticleConnection = {
	__typename?: "ArticleConnection"
	/** A list of edges. */
	edges: Array<ArticleEdge>
	/** A list of the nodes contained in ArticleEdge. */
	nodes: Array<Article>
	/** Information to aid in pagination. */
	pageInfo: PageInfo
}
// ...
export type Blog = HasMetafields &
	Node &
	OnlineStorePublishable & {
		__typename?: "Blog"
		/** Find an article by its handle. */
		articleByHandle?: Maybe<Article>
		/** List of the blog's articles. */
		articles: ArticleConnection
		/** The authors who have contributed to the blog. */
		authors: Array<ArticleAuthor>
		/**
		 * A human-friendly unique string for the Blog automatically generated from its title.
		 *
		 */
		handle: Scalars["String"]["output"]
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars["URL"]["output"]>
		/** The blog's SEO information. */
		seo?: Maybe<Seo>
		/** The blogs’s title. */
		title: Scalars["String"]["output"]
	}
// ...
export type CommentConnection = {
	__typename?: "CommentConnection"
	/** A list of edges. */
	edges: Array<CommentEdge>
	/** A list of the nodes contained in CommentEdge. */
	nodes: Array<Comment>
	/** Information to aid in pagination. */
	pageInfo: PageInfo
}
// ...
export type HasMetafields = {
	/** Returns a metafield found by namespace and key. */
	metafield?: Maybe<Metafield>
	/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
	metafields: Array<Maybe<Metafield>>
}
// ...
export type Node = {
	/** A globally-unique ID. */
	id: Scalars["ID"]["output"]
}

/** Represents a resource that can be published to the Online Store sales channel. */
export type OnlineStorePublishable = {
	/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
	onlineStoreUrl?: Maybe<Scalars["URL"]["output"]>
}
// ...
export type Seo = {
	__typename?: "SEO"
	/** The meta description. */
	description?: Maybe<Scalars["String"]["output"]>
	/** The SEO title. */
	title?: Maybe<Scalars["String"]["output"]>
}
// ...
export type Trackable = {
	/** A URL parameters to be added to a page URL when it is linked from a GraphQL result. This allows for tracking the origin of the traffic. */
	trackingParameters?: Maybe<Scalars["String"]["output"]>
}
// ...
export type ArticleByIdQueryVariables = Exact<{
	id: Scalars["ID"]["input"]
}>

export type ArticleByIdQuery = {
	__typename?: "QueryRoot"
	record?: {
		__typename?: "Article"
		id: string
		handle: string
		blog: { __typename?: "Blog"; title: string }
	} | null
}

export type ArticleByIdentifierQueryVariables = Exact<{
	first: Scalars["Int"]["input"]
	after?: InputMaybe<Scalars["String"]["input"]>
	query?: InputMaybe<Scalars["String"]["input"]>
}>

export type ArticleByIdentifierQuery = {
	__typename?: "QueryRoot"
	articles: {
		__typename?: "ArticleConnection"
		nodes: Array<{
			__typename?: "Article"
			id: string
			handle: string
			blog: { __typename?: "Blog"; title: string }
		}>
		pageInfo: {
			__typename?: "PageInfo"
			hasNextPage: boolean
			endCursor?: string | null
		}
	}
}

export type ArticleFragmentFragment = {
	__typename?: "Article"
	id: string
	handle: string
	blog: { __typename?: "Blog"; title: string }
}
// ...
export const ArticleByIdDocument = {
	kind: "Document",
	definitions: [
		{
			kind: "OperationDefinition",
			operation: "query",
			name: { kind: "Name", value: "articleById" },
			variableDefinitions: [
				{
					kind: "VariableDefinition",
					variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
					type: {
						kind: "NonNullType",
						type: { kind: "NamedType", name: { kind: "Name", value: "ID" } },
					},
				},
			],
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{
						kind: "Field",
						alias: { kind: "Name", value: "record" },
						name: { kind: "Name", value: "article" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "id" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "id" },
								},
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "FragmentSpread",
									name: { kind: "Name", value: "ArticleFragment" },
								},
							],
						},
					},
				],
			},
		},
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "ArticleFragment" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Article" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{ kind: "Field", name: { kind: "Name", value: "handle" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "blog" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [{ kind: "Field", name: { kind: "Name", value: "title" } }],
						},
					},
				],
			},
		},
	],
} as unknown as DocumentNode<ArticleByIdQuery, ArticleByIdQueryVariables>
export const ArticleByIdentifierDocument = {
	kind: "Document",
	definitions: [
		{
			kind: "OperationDefinition",
			operation: "query",
			name: { kind: "Name", value: "articleByIdentifier" },
			variableDefinitions: [
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "first" },
					},
					type: {
						kind: "NonNullType",
						type: { kind: "NamedType", name: { kind: "Name", value: "Int" } },
					},
				},
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "after" },
					},
					type: { kind: "NamedType", name: { kind: "Name", value: "String" } },
				},
				{
					kind: "VariableDefinition",
					variable: {
						kind: "Variable",
						name: { kind: "Name", value: "query" },
					},
					type: { kind: "NamedType", name: { kind: "Name", value: "String" } },
				},
			],
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{
						kind: "Field",
						name: { kind: "Name", value: "articles" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "first" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "first" },
								},
							},
							{
								kind: "Argument",
								name: { kind: "Name", value: "after" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "after" },
								},
							},
							{
								kind: "Argument",
								name: { kind: "Name", value: "query" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "query" },
								},
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "nodes" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "FragmentSpread",
												name: { kind: "Name", value: "ArticleFragment" },
											},
										],
									},
								},
								{
									kind: "Field",
									name: { kind: "Name", value: "pageInfo" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{
												kind: "Field",
												name: { kind: "Name", value: "hasNextPage" },
											},
											{
												kind: "Field",
												name: { kind: "Name", value: "endCursor" },
											},
										],
									},
								},
							],
						},
					},
				],
			},
		},
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "ArticleFragment" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Article" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{ kind: "Field", name: { kind: "Name", value: "handle" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "blog" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [{ kind: "Field", name: { kind: "Name", value: "title" } }],
						},
					},
				],
			},
		},
	],
} as unknown as DocumentNode<ArticleByIdentifierQuery, ArticleByIdentifierQueryVariables>